refactor(tests): extract integration test base
